application/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Flask==2.0.2
Flask-Login==0.6.2
Flask-RESTful==0.3.10
Flask-Security-Too==4.1.3
Flask-SQLAlchemy==2.5.1
pyrsistent==0.19.3
SQLAlchemy==1.4.25
Werkzeug==2.3.6

WARNING(pigar): some manual fixes are required since pigar has found duplicate requirements for the same import name.
# WARNING(pigar): the following duplicate requirements are for import name: application
ally-py==0.0.0
application==1.0
codb==1.2.0
cryptoasset-data-downloader==1.0.9
dalio==0.0.1
DepartmnetHelper==1.0.0
docker-compose-multistage==0.1
ecope==0.1.3
flask-wheel==0.0.4
PedalPi-Application==0.4.1
Pforests-dtw==1.3
Proximity-Forest-dtw==1.1
yangxian==1.1
